Separate MinerReward action
